requirements.txt
Flask
opencv-contrib-python
opencv-python
numpy
Pillow
